requirements.txt
pandas
numpy
tensorflow
keras
tensorboard
streamlit
pillow
requests
python-dotenv